+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Latest changes
 
+* VFS-13271 Fixed selecting invite token target in sidebar filter when there is a token for non-existing entity
 * VFS-13560 Added new development scripts
 * VFS-13712 Fixed live refresh for spaces list in harvesters
 * VFS-13461 Added provider version on providers page